review modif: move loop flow flag to rao parameter
