+++ ra-optimisation/linear-rao/src/main/java/com/farao_community/farao/linear_rao/LinearRao.java
@@ -11,13 +11,11 @@ import com.farao_community.farao.commons.FaraoException;
 import com.farao_community.farao.data.crac_api.Cnec;
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_api.Unit;
-import com.farao_community.farao.data.crac_loopflow_extension.CracLoopFlowExtension;
 import com.farao_community.farao.linear_rao.config.LinearRaoConfigurationUtil;
 import com.farao_community.farao.linear_rao.config.LinearRaoParameters;
 import com.farao_community.farao.ra_optimisation.*;
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.farao_community.farao.rao_api.RaoProvider;
-import com.farao_community.farao.search_tree_rao.config.SearchTreeRaoParameters;
 import com.farao_community.farao.util.SystematicSensitivityAnalysisResult;
 import com.farao_community.farao.util.SystematicSensitivityAnalysisService;
 import com.google.auto.service.AutoService;
@@ -82,12 +80,7 @@ public class LinearRao implements RaoProvider {
 
         SystematicSensitivityAnalysisResult tempSensitivityAnalysisResult;
 
-        //flag for using loopflow constraint, pass to LinearRaoModeller
-        SearchTreeRaoParameters searchTreeRaoParameters = parameters.getExtension(SearchTreeRaoParameters.class);
-        CracLoopFlowExtension cracLoopFlowExtension = crac.getExtension(CracLoopFlowExtension.class);
-        boolean useLoopFlow = !Objects.isNull(searchTreeRaoParameters) && searchTreeRaoParameters.isRaoWithLoopFlow() && !Objects.isNull(cracLoopFlowExtension);
-
-        LinearRaoModeller linearRaoModeller = createLinearRaoModeller(crac, network, preOptimSensitivityAnalysisResult, useLoopFlow);
+        LinearRaoModeller linearRaoModeller = createLinearRaoModeller(crac, network, preOptimSensitivityAnalysisResult, parameters);
         linearRaoModeller.buildProblem();
 
         RaoComputationResult raoComputationResult;
@@ -135,8 +128,8 @@ public class LinearRao implements RaoProvider {
     LinearRaoModeller createLinearRaoModeller(Crac crac,
                                               Network network,
                                               SystematicSensitivityAnalysisResult systematicSensitivityAnalysisResult,
-                                              boolean useLoopFlow) {
-        return new LinearRaoModeller(crac, network, systematicSensitivityAnalysisResult, new LinearRaoProblem(), useLoopFlow);
+                                              RaoParameters raoParameters) {
+        return new LinearRaoModeller(crac, network, systematicSensitivityAnalysisResult, new LinearRaoProblem(), raoParameters);
     }
 
     private double getRemedialActionResultPostOptimisationValue(RemedialActionResult remedialActionResult) {
