+++ ra-optimisation/linear-rao/src/test/resources/LinearRaoParameters.json
@@ -2,6 +2,7 @@
   "version" : "1.0",
   "dc-mode" : false,
   "ac-to-dc-fallback" : false,
+  "rao-with-loop-flow" : false,
   "extensions" : {
     "LinearRaoParameters" : {
       "max-number-of-iterations" : 20,
