+++ ra-optimisation/rao-api/src/main/java/com/farao_community/farao/rao_api/json/RaoParametersSerializer.java
@@ -32,6 +32,7 @@ public class RaoParametersSerializer extends StdSerializer<RaoParameters> {
 
         jsonGenerator.writeBooleanField("dc-mode", parameters.isDcMode());
         jsonGenerator.writeBooleanField("ac-to-dc-fallback", parameters.isAcToDcFallback());
+        jsonGenerator.writeBooleanField("rao-with-loop-flow", parameters.isRaoWithLoopFlow());
 
         JsonUtil.writeExtensions(parameters, jsonGenerator, serializerProvider, JsonRaoParameters.getExtensionSerializers());
