+++ ra-optimisation/rao-api/src/test/resources/RaoParameters.json
@@ -1,5 +1,6 @@
 {
   "version" : "1.0",
   "dc-mode" : true,
-  "ac-to-dc-fallback" : true
+  "ac-to-dc-fallback" : true,
+  "rao-with-loop-flow" : false
 }
