+++ ra-optimisation/rao-api/src/test/resources/RaoParametersWithExtension.json
@@ -2,6 +2,7 @@
   "version" : "1.0",
   "dc-mode" : true,
   "ac-to-dc-fallback" : true,
+  "rao-with-loop-flow" : false,
   "extensions" : {
     "dummy-extension" : { }
   }
