+++ ra-optimisation/search-tree-rao/src/test/java/com.farao_community.farao.search_tree_rao/config/SearchTreeRaoParametersTest.java
@@ -49,7 +49,5 @@ public class SearchTreeRaoParametersTest {
     public void testLoopFlowExtensionParameters() {
         parameters.addExtension(SearchTreeRaoParameters.class, new SearchTreeRaoParameters());
         SearchTreeRaoParameters searchTreeRaoParameters = parameters.getExtension(SearchTreeRaoParameters.class);
-        searchTreeRaoParameters.setRaoWithLoopFlow(true);
-        assertTrue(searchTreeRaoParameters.isRaoWithLoopFlow());
     }
 }
